Moving Clusterizer.h to include
